backend/requirements.txt
# Django Framework
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.11

# Configuration
python-decouple==3.8

# Data Processing
pandas==3.0.0
openpyxl==3.1.5
numpy==2.4.2

# Image Processing
pillow==12.1.0

# Google Sheets Integration
gspread==6.2.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0

# HTTP Requests
requests==2.32.5
requests-oauthlib==2.0.0

# Dependencies
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
et-xmlfile==2.0.0
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
